Restructure project file tree
Further refactoring to adhere to the MVC design pattern
