+++ docs/Javadocs/serialized-form.html
@@ -76,8 +76,8 @@ loadScripts(document, 'script');</script>
 <pre><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/javax/swing/JFrame.html" title="class or interface in javax.swing" class="external-link">JFrame</a> owner</pre>
 </li>
 <li class="block-list">
-<h5>com.hcyic1.level</h5>
-<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.level</pre>
+<h5>com.hcyic1.game</h5>
+<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.game</pre>
 </li>
 </ul>
 </section>
@@ -111,8 +111,8 @@ loadScripts(document, 'script');</script>
 <pre><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/javax/swing/JButton.html" title="class or interface in javax.swing" class="external-link">JButton</a> skipLevel</pre>
 </li>
 <li class="block-list">
-<h5>com.hcyic1.level</h5>
-<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.level</pre>
+<h5>com.hcyic1.game</h5>
+<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.game</pre>
 </li>
 </ul>
 </section>
@@ -166,8 +166,8 @@ loadScripts(document, 'script');</script>
 <pre>int strLen</pre>
 </li>
 <li class="block-list">
-<h5>com.hcyic1.level</h5>
-<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.level</pre>
+<h5>com.hcyic1.game</h5>
+<pre><a href="test/Wall.html" title="class in test">Wall</a> com.hcyic1.game</pre>
 </li>
 </ul>
 </section>
