+++ docs/Javadocs/test/DebugConsole.html
@@ -154,7 +154,7 @@ implements <a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desk
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
 <div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(javax.swing.JFrame,test.Wall,test.GameBoard)" class="member-name-link">DebugConsole</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/javax/swing/JFrame.html" title="class or interface in javax.swing" class="external-link">JFrame</a>&nbsp;owner,
- <a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.level,
+ <a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.game,
  <a href="GameBoard.html" title="class in test">GameBoard</a>&nbsp;gameBoard)</code></div>
 <div class="col-last even-row-color">&nbsp;</div>
 </div>
@@ -228,7 +228,7 @@ implements <a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desk
 <section class="detail" id="&lt;init&gt;(javax.swing.JFrame,test.Wall,test.GameBoard)">
 <h3>DebugConsole</h3>
 <div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">DebugConsole</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/javax/swing/JFrame.html" title="class or interface in javax.swing" class="external-link">JFrame</a>&nbsp;owner,
- <a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.level,
+ <a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.game,
  <a href="GameBoard.html" title="class in test">GameBoard</a>&nbsp;gameBoard)</span></div>
 </section>
 </li>
