+++ docs/Javadocs/test/DebugPanel.html
@@ -141,7 +141,7 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
-<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(test.Wall)" class="member-name-link">DebugPanel</a><wbr>(<a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.level)</code></div>
+<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(test.Wall)" class="member-name-link">DebugPanel</a><wbr>(<a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.game)</code></div>
 <div class="col-last even-row-color">&nbsp;</div>
 </div>
 </section>
@@ -193,7 +193,7 @@ loadScripts(document, 'script');</script>
 <li>
 <section class="detail" id="&lt;init&gt;(test.Wall)">
 <h3>DebugPanel</h3>
-<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">DebugPanel</span><wbr><span class="parameters">(<a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.level)</span></div>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">DebugPanel</span><wbr><span class="parameters">(<a href="Wall.html" title="class in test">Wall</a>&nbsp;com.hcyic1.game)</span></div>
 </section>
 </li>
 </ul>
