+++ src/main/java/com/hcyic1/model/ball/RubberBall.java
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hcyic1.ball;
+package com.hcyic1.model.ball;
 
 import java.awt.*;
 import java.awt.geom.Ellipse2D;
