+++ src/main/java/com/hcyic1/model/brick/SteelBrick.java
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hcyic1.brick;
+package com.hcyic1.model.brick;
 
 import java.awt.*;
 import java.awt.geom.Point2D;
