+++ src/main/java/com/hcyic1/model/game/GameBoard.java
@@ -15,13 +15,14 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hcyic1.gui;
+package com.hcyic1.model.game;
 
-import com.hcyic1.ball.Ball;
-import com.hcyic1.brick.Brick;
-import com.hcyic1.debug.DebugConsole;
-import com.hcyic1.platform.Platform;
-import com.hcyic1.level.Level;
+import com.hcyic1.highscore.HighScoreFile;
+import com.hcyic1.model.ball.Ball;
+import com.hcyic1.model.brick.Brick;
+import com.hcyic1.view.DebugConsole;
+import com.hcyic1.model.platform.Platform;
+import com.hcyic1.highscore.HighScoreView;
 
 import javax.swing.*;
 import java.awt.*;
@@ -68,6 +69,8 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
 
     private DebugConsole debugConsole;
 
+    public HighScoreFile highScoreFile = new HighScoreFile();
+    public HighScoreView highScoreView = new HighScoreView(highScoreFile);
 
     public GameBoard(JFrame owner) {
         super();
@@ -87,6 +90,9 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
         //initialize the first level
         level.nextLevel();
 
+        // high score intiialisation
+        highScoreFile.addOrUpdateScore(level.player);
+
         gameTimer = new Timer(10, e -> {
             level.move();
             level.findImpacts();
@@ -98,21 +104,28 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
                     level.player.getScoreMultiplier()
                     );
             if (level.isBallLost()) {
+                highScoreFile.addOrUpdateScore(level.player);
                 if (level.ballEnd()) {
                     level.wallReset();
                     message = "Game over";
+                    highScoreView.showScoreTable();
                 }
                 level.ballReset();
                 gameTimer.stop();
             } else if (level.isDone()) {
                 if (level.hasLevel()) {
                     message = "Go to Next Level";
+                    // show high score screen
+                    highScoreFile.addOrUpdateScore(level.player);
+                    highScoreView.showScoreTable();
                     gameTimer.stop();
                     level.ballReset();
                     level.wallReset();
                     level.nextLevel();
                 } else {
                     message = "ALL WALLS DESTROYED";
+                    highScoreFile.addOrUpdateScore(level.player);
+                    highScoreView.showScoreTable();
                     gameTimer.stop();
                 }
             }
@@ -231,18 +244,10 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
         g2d.setFont(menuFont);
         g2d.setColor(MENU_COLOR);
 
-        if (strLen == 0) {
-            FontRenderContext frc = g2d.getFontRenderContext();
-            strLen = menuFont.getStringBounds(PAUSE, frc).getBounds().width;
-        }
+        drawTitleString(g2d, PAUSE);
 
-        int x = (this.getWidth() - strLen) / 2;
-        int y = this.getHeight() / 10;
-
-        g2d.drawString(PAUSE, x, y);
-
-        x = this.getWidth() / 8;
-        y = this.getHeight() / 4;
+        int x = this.getWidth() / 8;
+        int y = this.getHeight() / 4;
 
 
         if (continueButtonRect == null) {
@@ -276,6 +281,18 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
         g2d.setColor(tmpColor);
     }
 
+    private void drawTitleString(Graphics2D g2d, String string) {
+        if (strLen == 0) {
+            FontRenderContext frc = g2d.getFontRenderContext();
+            strLen = menuFont.getStringBounds(string, frc).getBounds().width;
+        }
+
+        int x = (this.getWidth() - strLen) / 2;
+        int y = this.getHeight() / 10;
+
+        g2d.drawString(string, x, y);
+    }
+
     @Override
     public void keyTyped(KeyEvent keyEvent) {
     }
@@ -304,6 +321,10 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
             case KeyEvent.VK_F1:
                 if (keyEvent.isAltDown() && keyEvent.isShiftDown())
                     debugConsole.setVisible(true);
+            case KeyEvent.VK_H:
+                // show high score screen
+                highScoreFile.addOrUpdateScore(level.player);
+                highScoreView.showScoreTable();
             default:
                 level.platform.stop();
         }
@@ -326,6 +347,7 @@ public class GameBoard extends JComponent implements KeyListener, MouseListener,
             message = "Restarting Game...";
             level.ballReset();
             level.wallReset();
+            // need to reset scores
             showPauseMenu = false;
             repaint();
         } else if (exitButtonRect.contains(p)) {
