+++ src/main/java/com/hcyic1/view/DebugPanel.java
@@ -15,9 +15,9 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hcyic1.debug;
+package com.hcyic1.view;
 
-import com.hcyic1.level.Level;
+import com.hcyic1.model.game.Level;
 
 import javax.swing.*;
 import javax.swing.event.ChangeListener;
